requirements.txt
streamlit==1.39.0
streamlit-drawable-canvas==0.9.3
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
scikit-image==0.24.0
tensorflow==2.16.1
